feat: build all 4 targets, fix compilation errors
- Linux x86_64 native build: WORKING (2.5MB ELF executable)
- Windows x86_64: partial (MinGW lacks complete sysroot in .local)
- macOS x86_64: partial (needs cross-compiled GMP/MPFR for target)
- JAR: WORKING (115KB with embedded libcalc-bridge.so)

Fixed compilation errors:
- Added #include <mpfr.h> to big_decimal.h (was using mpfr_t without header)
- Added #include <atomic> to big_decimal.cpp
- Created expression_parser.h public header, refactored expression_parser.cpp
- Fixed BoundType mismatch between expression.h and set_operations.h
- Fixed CalcSet non-copyable issue (added clone() method)
- Fixed BigDecimal(int, int) ambiguity with int64_t(1) cast
- Added detectBase() to base_converter.cpp
- Added #include <cstdint>/<memory> to font_manager.h
- Added #include <string> to fonts.h
- Fixed ImGui InputText to use char buffers instead of std::string*
- Fixed namespace nesting in fonts.h (font_latinmodern_math.h include)
- Removed 16-bit-overflow ImWchar ranges in main.cpp
- Added CalcEngine::~CalcEngine() declaration in header
- Added #include <stdexcept> to crypto_random.cpp

Build infrastructure:
- build.sh script for all 4 targets
- Cross-compile toolchains: MinGW-w64 (Windows), Zig (macOS)
- Dear ImGui downloaded to third_party/imgui/
- All dependencies resolved: GMP, MPFR, GLFW, OpenGL
- .local/ directory excluded from git (cross-compile toolchain)

+++ calc-core/src/main/cpp/src/functions.cpp
@@ -161,7 +161,7 @@ void FunctionRegistry::registerBuiltins() {
                 return args[0].nthRoot(2);
             } else {
                 // a-th root of b = b^(1/a)
-                BigDecimal reciprocal = BigDecimal(1, args[0].getPrecision()).div(args[0]);
+                BigDecimal reciprocal = BigDecimal(int64_t(1), args[0].getPrecision()).div(args[0]);
                 return args[1].pow(reciprocal);
             }
         },
@@ -243,7 +243,7 @@ void FunctionRegistry::registerBuiltins() {
                 [](mpfr_t result, const mpfr_t arg, mpfr_rnd_t rnd) {
                     mpfr_cos(result, arg, rnd);
                 });
-            return BigDecimal(1, args[0].getPrecision()).div(cosVal);
+            return BigDecimal(int64_t(1), args[0].getPrecision()).div(cosVal);
         },
         RenderSpec{RenderSpec::Type::PARENTHESIS, "sec"},
         1, 1
@@ -256,7 +256,7 @@ void FunctionRegistry::registerBuiltins() {
                 [](mpfr_t result, const mpfr_t arg, mpfr_rnd_t rnd) {
                     mpfr_sin(result, arg, rnd);
                 });
-            return BigDecimal(1, args[0].getPrecision()).div(sinVal);
+            return BigDecimal(int64_t(1), args[0].getPrecision()).div(sinVal);
         },
         RenderSpec{RenderSpec::Type::PARENTHESIS, "csc"},
         1, 1
@@ -301,7 +301,7 @@ void FunctionRegistry::registerBuiltins() {
     // arccot = arctan(1/x)
     builtins_["arccot"] = BuiltinEntry{
         [](const std::vector<BigDecimal>& args) -> BigDecimal {
-            BigDecimal recip = BigDecimal(1, args[0].getPrecision()).div(args[0]);
+            BigDecimal recip = BigDecimal(int64_t(1), args[0].getPrecision()).div(args[0]);
             return mpfrUnaryOp(recip, recip.getPrecision(),
                 [](mpfr_t result, const mpfr_t arg, mpfr_rnd_t rnd) {
                     mpfr_atan(result, arg, rnd);
@@ -314,7 +314,7 @@ void FunctionRegistry::registerBuiltins() {
     // arcsec = arccos(1/x)
     builtins_["arcsec"] = BuiltinEntry{
         [](const std::vector<BigDecimal>& args) -> BigDecimal {
-            BigDecimal recip = BigDecimal(1, args[0].getPrecision()).div(args[0]);
+            BigDecimal recip = BigDecimal(int64_t(1), args[0].getPrecision()).div(args[0]);
             return mpfrUnaryOp(recip, recip.getPrecision(),
                 [](mpfr_t result, const mpfr_t arg, mpfr_rnd_t rnd) {
                     mpfr_acos(result, arg, rnd);
@@ -327,7 +327,7 @@ void FunctionRegistry::registerBuiltins() {
     // arccsc = arcsin(1/x)
     builtins_["arccsc"] = BuiltinEntry{
         [](const std::vector<BigDecimal>& args) -> BigDecimal {
-            BigDecimal recip = BigDecimal(1, args[0].getPrecision()).div(args[0]);
+            BigDecimal recip = BigDecimal(int64_t(1), args[0].getPrecision()).div(args[0]);
             return mpfrUnaryOp(recip, recip.getPrecision(),
                 [](mpfr_t result, const mpfr_t arg, mpfr_rnd_t rnd) {
                     mpfr_asin(result, arg, rnd);
